feat(foundation): implement Span 8-byte struct
Closes #1

- Create dedicated span.rs module with Span struct
- Use u32 instead of usize for start/end (8 bytes vs 16)
- Add len(), contains(), is_empty(), merge(), dummy() methods
- Add #[repr(C)] for predictable memory layout
- Include compile-time size assertion
- Update lexer and error reporting to cast usize to u32
- Re-export Span from token.rs for backward compatibility

Tests verify:
- Size is exactly 8 bytes (ADR-002 compliance)
- All methods work correctly
- Hash consistency for use in collections

+++ bootstrap/src/error.rs
@@ -146,11 +146,11 @@ pub fn report(source: &str, filename: &str, error: NovaError) {
     let message = error.message();
     let code = error.code();
 
-    Report::build(ReportKind::Error, filename, span.start)
+    Report::build(ReportKind::Error, filename, span.start as usize)
         .with_code(code)
         .with_message(&message)
         .with_label(
-            Label::new((filename, span.start..span.end))
+            Label::new((filename, span.start as usize..span.end as usize))
                 .with_message(&message)
                 .with_color(Color::Red),
         )
